# Optimize NPB-FT evolve() by using localAccess